update manual workaround to be more robust to future changes

+++ .github/scripts/psycopg2-check.sh
@@ -3,9 +3,11 @@ source venv/bin/activate
 python -m pip install .
 
 if [[ "$PSYCOPG2_WORKAROUND" == true ]]; then
-    PSYCOPG2_VERSION=$(pip show psycopg2-binary | grep Version | cut -d " " -f 2)
-    pip uninstall -y psycopg2-binary
-    pip install psycopg2==$PSYCOPG2_VERSION
+    if [[ pip show psycopg2-binary ]]; then
+        PSYCOPG2_VERSION=$(pip show psycopg2-binary | grep Version | cut -d " " -f 2)
+        pip uninstall -y psycopg2-binary
+        pip install psycopg2==$PSYCOPG2_VERSION
+    fi
 fi
 
 PSYCOPG2_NAME=$((pip show psycopg2 || pip show psycopg2-binary) | grep Name | cut -d " " -f 2)
